refactor: 移除不必要的构造函数调用并添加默认空动作实现
