+++ TeamCode/src/main/java/org/betastudio/ftc/action/utils/NullptrAction.java
@@ -1,0 +1,15 @@
+package org.betastudio.ftc.action.utils;
+
+import org.betastudio.ftc.action.Action;
+
+public class NullptrAction implements Action {
+	@Override
+	public boolean activate() {
+		return false;
+	}
+
+	@Override
+	public String paramsString() {
+		return "null";
+	}
+}
