+++ TeamCode/src/main/java/org/betastudio/ftc/action/utils/SleepingAction.java
@@ -2,6 +2,7 @@ package org.betastudio.ftc.action.utils;
 
 import androidx.annotation.NonNull;
 
+import org.betastudio.ftc.action.ActionImpl;
 import org.betastudio.ftc.util.time.Timer;
 import org.jetbrains.annotations.Contract;
 
@@ -15,9 +16,6 @@ public final class SleepingAction extends ActionImpl {
 	private final AtomicBoolean                      initialized = new AtomicBoolean(false);
 
 	public SleepingAction(final long sleepMilliseconds) {
-		// 先调用父类构造函数，传入默认值 null
-		super(()->true);
-		// 设置 actionRef 的值
 		final AtomicReference <Callable <Boolean>> actionRef = new AtomicReference <>();
 		actionRef.set(() -> Timer.getCurrentTime() - startTime.get() >= sleepMilliseconds);
 		final Callable <Boolean> action = actionRef.get();
